.history/requirements_20250203072942.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
certifi==2025.1.31
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==1.61.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.3.4
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
